Make childViewEventHandler a common module
